6 commit, it`s all pointless

+++ src/main/java/com/example/fitness/repositories/MembershipsRepository.java
@@ -3,5 +3,5 @@ package com.example.fitness.repositories;
 import com.example.fitness.domain.Memberships;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MembershipsRepository extends CrudRepository<Memberships, Long> {
+public interface MembershipsRepository extends CrudRepository<Memberships, Integer> {
 }
