+++ src/main/java/com/example/fitness/repositories/UsersRepository.java
@@ -3,5 +3,5 @@ package com.example.fitness.repositories;
 import com.example.fitness.domain.Users;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsersRepository extends CrudRepository<Users, Long> {
+public interface UsersRepository extends CrudRepository<Users, Integer> {
 }
